Move timestamp hash test from hash index test to timestamp basic test
